requirements.txt
backtrader
pyfolio
yfinance
ta 
python-dotenv
yfinance
alpaca-py
pandas
pytest
black
mypy
ruff
pytest-cov
pluggy
pre-commit


